uncomment

+++ scripts/run_integration_tests.sh
@@ -18,12 +18,18 @@ until curl --insecure --silent --fail "$COSMOSDB_CERT_URL"; do
   retry_count=$((retry_count+1))
 done
 
-echo "Cosmos DB Emulator is healthy ..."
+echo "Cosmos DB Emulator is healthy"
 
-echo 'Outputting certificate to host certificate store ...'
+if [ ! -f /usr/local/share/ca-certificates/cosmos-db-emulator.crt ]; then
+  echo "Failed to download Cosmos DB certificate."
+  exit 1
+fi
+
+echo 'CosmosDb certificate - move to host certificate store ...'
 
 curl -k "$COSMOSDB_CERT_URL" -o /usr/local/share/ca-certificates/cosmos-db-emulator.crt
 
+echo 'CosmosDb Certificate done'
 
 # Resolve absolute path to the directory containing this script
 
@@ -31,12 +37,11 @@ SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
 ## WireMock certificate
 
-WIREMOCK_SCRIPT_PATH="$SCRIPT_DIR/generate_wiremock_server_cert.sh"
-chmod +x "$WIREMOCK_SCRIPT_PATH" && /bin/bash -c $WIREMOCK_SCRIPT_PATH
+"$SCRIPT_DIR/generate_wiremock_server_cert.sh"
 
 update-ca-certificates
 
 # Run integration tests using absolute path
- dotnet test "$SCRIPT_DIR/../DfE.GIAP.All/tests/DfE.GIAP.Core.IntegrationTests/DfE.GIAP.Core.IntegrationTests.csproj" --no-build --no-restore --nologo --logger "console;verbosity=detailed"
+ dotnet test "$SCRIPT_DIR/../DfE.GIAP.All/tests/DfE.GIAP.Core.IntegrationTests/DfE.GIAP.Core.IntegrationTests.csproj" --no-build --no-restore --nologo --logger "console;verbosity=detailed" -p:CollectCoverage=true -p:CoverletOutputFormat=cobertura -p:CoverletOutput="$SCRIPT_DIR/../DfE.GIAP.All/coverlet/reports/DfE.GIAP.Core.IntegrationTests/cobertura.xml"
 
 #tail -f /dev/null
